Experiments

Documentation for Experiments.

This package defines a framework for generating useful data in a consistent manner.

The idea is that, for any given experiment, you can implement a few intuitive structs organizing your input parameters, and a few intuitive methods to actually run your experiment, and this class handles the IO (ie. writing to a CSV file).

This framework is only particularly useful when the outputs you really care about are all scalar. I think it's good practice to always try boiling a problem down to scalar inputs and outputs when you can, but it won't always work. Personally, I see this framework as useful for late-stage research: I have already figured out what the trends are, I know what I want to plot, and I can use this package to help me generate a lot of data.

Experiments.Experiments

Experiments.IndexType

Experiments.collectdata

Experiments.collectdata

Experiments.header

Experiments.record

Experiments.runtrial

Experiments.synthesize

Experiments.Experiments — Module

A unique collection of variables and instructions to generate useful data.

Implementation

Defining a new experiment requires implementing five structs and four methods.

The structs to be implemented are:

Control contains all control variables, and serves as primary dispatch object

Setup contains all complex objects calculable from just control variables

Independent contains all independent variables (ie. stuff you plan to plot against)

Result contains all complex objects calculated from independent variables

Dependent contains all dependent variables (ie. stuff you plan to plot)

IMPORTANT:

One of the main functions of Control, Independent, and Dependent is to be written to a csv file. Thus, all attributes in these structs must be directly serializable without tabs or newlines.

The methods to be implemented are:

initialize(::Control)::Setup

mapindex(::Control, ::Integer)::Independent

This maps an integer index to a specific selection of independent variables, facilitating en masse data collection.

runtrial(::Control, ::Setup, ::Independent)::Result

synthesize(::Control, ::Setup, ::Independent, ::Result)::Dependent

By implementing these methods, several additional methods are automatically available. The most important of these is collectdata, so find its documentation too!

Experiments.IndexType — Type

IndexType

The index argument in collectdata can in principle be any iterable, but for the sake of proper dispatch we seem to need to specify it.

Experiments.collectdata — Method

collectdata(io::IO, expmt::Control, setup::Setup, index; writeheader=false)

Run a series of trials and write the results in tab-delimited .csv format. For convenience, return the very last result.

index is an iterable of Integer; each is passed to mapindex to generate a collection of Independent variables.

Pass writeheader=true to include a header line in the resulting .csv file.

Experiments.collectdata — Method

collectdata(io::IO, expmt::Control, index; writeheader=false)

Initialize experiment from scratch before collecting data.

Experiments.header — Method

header(io::IO, expmt::Control, xvars::Independent, yvars::Dependent)

Write all attribute names to io, separated by tabs.

Attributes in Independent are written first, then Dependent, then Control, in the order that they are defined in these structs.

Experiments.record — Method

record(io::IO, expmt::Control, xvars::Independent, yvars::Dependent)

Write all variables to io, separated by tabs.

Variables in Independent are written first, then Dependent, then Control, in the order that they are defined in these structs.

Experiments.runtrial — Method

runtrial(expmt::Control, xvars::Independent)

Initialize experiment from scratch before running trial.

Experiments.synthesize — Method

synthesize(expmt::Control, xvars::Independent)

Initialize and run experiment from scratch before synthesizing result.