Experiments

Documentation for Experiments.

Experiments.INERT

Experiments.Q_CO2

Experiments.Q_O2

Experiments.Q_S

Experiments.RQ

Experiments.duration

Experiments.duration

Experiments.endtime

Experiments.endtime

Experiments.starttime

Experiments.starttime

Experiments.x_CGin

Experiments.x_OGin

Experiments.INERT — Method

INERT(xOGin,xCGin,xO2,xCO2;EXH2O=EXH2O()) = (1 .-xOGin.-xCGin)./(1 .-xO2.-xCO2.-EXH2O)

Experiments.Q_CO2 — Method

QCO2(FAIR,FO2,xCO2,xCGin,INERT;VM=22.414) = (FAIR+FO2)./VM.*(xCO2.*INERT-x_CGin)

Experiments.Q_O2 — Method

QO2(FAIR,FO2,xO2,xOGin,INERT;VM=22.414) = (FAIR+FO2)./VM.*(xO2.*INERT-x_OGin)

Experiments.Q_S — Method

QS(FR;cSR=400,MS=30) = -FR .* cSR ./ M_S

Experiments.RQ — Method

RQ(QCO2,QO2) = -QO2 ./ QCO2

Experiments.duration — Method

duration(experiment::Experiment)

Get the duration of an experiment from all included timeseries.

Examples

julia> duration(experiment)
1 day, 10 hours, 9 minutes, 55 seconds

Experiments.duration — Method

duration(timeseries::TimeArray)

Get the duration of a timeseries.

Examples

julia> duration(timeseries)
1 day, 10 hours, 9 minutes, 55 seconds

Experiments.endtime — Method

endtime(experiment::Experiment)

Get the end datetime of an experiment from the maximum of all included timeseries.

Examples

julia> endtime(experiment)
2019-01-02T10:10:00

Experiments.endtime — Method

endtime(timeseries::TimeArray)

Get the end datetime of a timeseries.

Examples

julia> endtime(timeseries)
2019-01-02T10:10:00

Experiments.starttime — Method

starttime(experiment::Experiment)

Get the start datetime of an experiment from the minimum of all included timeseries.

Examples

julia> starttime(experiment)
2019-01-01T00:00:00

Experiments.starttime — Method

starttime(timeseries::TimeArray)

Get the start datetime of a timeseries.

Examples

julia> starttime(timeseries)
2019-01-01T00:00:00

Experiments.x_CGin — Method

xCGin(FAIR,FO2;xCAIR=0.0004) = (FAIR.*xCAIR)./(FAIR+FO2)

Experiments.x_OGin — Method

xOGin(FAIR,FO2;xOAIR=0.2094) = (FAIR.*xOAIR+FO2)./(FAIR+F_O2)